requirements.txt
streamlit
openai
faiss-cpu
python-dotenv
pypdf